Added Simple Menu and Highscore screens

+++ Asteroids/RaylibStarterCPP/GameObject.cpp
@@ -17,17 +17,17 @@ void GameObject::Update()
 	pos.y += vel.y;
 
 	// Allows the game object to wrap around the screen and appear on the other side if it goes off screen.
-	if (pos.x > 600 + img.height / 2 * scale) {
-		pos.x -= (600 + img.width * scale);
+	if (pos.x > 600 + (float)img.height / 2 * scale) {
+		pos.x -= (600 + (float)img.width * scale);
 	}
-	if (pos.x < -img.height / 2 * scale) {
-		pos.x += (600 + img.width * scale);
+	if (pos.x < (float)-img.height / 2 * scale) {
+		pos.x += (600 + (float)img.width * scale);
 	}
-	if (pos.y > 600 + img.height / 2 * scale) {
-		pos.y -= (600 + img.width * scale);
+	if (pos.y > 600 + (float)img.height / 2 * scale) {
+		pos.y -= (600 + (float)img.width * scale);
 	}
-	if (pos.y < -img.height / 2 * scale) {
-		pos.y += (600 + img.width * scale);
+	if (pos.y < (float)-img.height / 2 * scale) {
+		pos.y += (600 + (float)img.width * scale);
 	}
 }
 
@@ -36,11 +36,11 @@ Vector2 GameObject::DrawOffset()
 	float moveDist = img.width * scale / 2;
 
 	// Moves the position of the object to be centred along the top edge.
-	Vector2 centredVec = { pos.x - cos((rotation)*DEG2RAD) * moveDist, pos.y + sin((rotation)*DEG2RAD) * moveDist };
+	Vector2 centredVec = { pos.x - (float)cos((rotation)*DEG2RAD) * moveDist, pos.y + (float)sin((rotation)*DEG2RAD) * moveDist };
 
 	// Moves the position of the object to be centred along the side edge.
-	centredVec.x -= sin((rotation)*DEG2RAD) * moveDist;
-	centredVec.y -= cos((rotation)*DEG2RAD) * moveDist;
+	centredVec.x -= (float)sin((rotation)*DEG2RAD) * moveDist;
+	centredVec.y -= (float)cos((rotation)*DEG2RAD) * moveDist;
 
 	return centredVec;
 }
